+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/internal/auth/CookieTokenRetriever.java
@@ -12,7 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.sso.token.auth;
+package com.liferay.portal.sso.token.internal.auth;
+
+import com.liferay.portal.kernel.util.CookieKeys;
+import com.liferay.portal.sso.token.auth.TokenLocation;
+import com.liferay.portal.sso.token.auth.TokenRetriever;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -22,18 +26,18 @@ import org.osgi.service.component.annotations.Component;
  * @author Michael C. Han
  */
 @Component(immediate = true, service = TokenRetriever.class)
-public class RequestHeaderTokenRetriever implements TokenRetriever {
+public class CookieTokenRetriever implements TokenRetriever {
 
 	@Override
 	public String getLoginToken(
 		HttpServletRequest request, String userTokenName) {
 
-		return request.getHeader(userTokenName);
+		return CookieKeys.getCookie(request, userTokenName, false);
 	}
 
 	@Override
 	public TokenLocation getTokenLocation() {
-		return TokenLocation.REQUEST_HEADER;
+		return TokenLocation.COOKIE;
 	}
 
 }
