+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/internal/auth/RequestTokenRetriever.java
@@ -12,9 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.sso.token.auth;
+package com.liferay.portal.sso.token.internal.auth;
 
-import com.liferay.portal.kernel.util.CookieKeys;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.sso.token.auth.TokenLocation;
+import com.liferay.portal.sso.token.auth.TokenRetriever;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -24,18 +26,18 @@ import org.osgi.service.component.annotations.Component;
  * @author Michael C. Han
  */
 @Component(immediate = true, service = TokenRetriever.class)
-public class CookieTokenRetriever implements TokenRetriever {
+public class RequestTokenRetriever implements TokenRetriever {
 
 	@Override
 	public String getLoginToken(
 		HttpServletRequest request, String userTokenName) {
 
-		return CookieKeys.getCookie(request, userTokenName, false);
+		return ParamUtil.getString(request, userTokenName);
 	}
 
 	@Override
 	public TokenLocation getTokenLocation() {
-		return TokenLocation.COOKIE;
+		return TokenLocation.REQUEST;
 	}
 
 }
